Add config interval option to snap to grid center
I overloaded the `interval` param to also accept `'corner'` and `'center'` to optionally instead snap to the corner of a grid square or the center of a grid square. (`'corner'` behaves exactly as setting interval to `1`. It's included just for completeness since it makes sense as an option knowing that `'center'` is also an option).

Getting the center is done by essentially using a snap of `2` and then just adding space to any snaps that are on a grid line. So the modulo math is just checking if the returned snap is directly on a grid line, if it is, then it's just adding a half a grid size to the x/y value on the snap object--thus every snapped object returned is guaranteed to be in the middle of a square.

p.s. I didn't update the readme or any of the inline comments here, so if something is out of sync those should be included. I also just copied the changes I made locally into github's editor so if it doesn't work then I missed something--but it's pretty straightforward so I'm pretty sure I grabbed it all (if I didn't feel free to ping me, but it's simple enough it'll probably be obvious what's missing in that case)